:recycle: ファイル分割
